app/backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.26.13
botocore==1.29.13
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
filelock==3.8.2
idna==3.4
jmespath==1.0.1
platformdirs==2.6.0
psycopg2-binary==2.9.4
python-dateutil==2.8.2
pytz==2022.5
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
uritools==4.0.0
urlextract==1.7.1
urllib3==1.26.12
requests
